requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
altair=5.1.1=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
asn1crypto=1.5.1=pyhd8ed1ab_0
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-auth=0.7.3=hab8b942_3
aws-c-cal=0.6.2=hb406d48_0
aws-c-common=0.9.0=hb547adb_0
aws-c-compression=0.2.17=he70778a_2
aws-c-event-stream=0.3.2=hcf14f3f_0
aws-c-http=0.7.12=he297698_1
aws-c-io=0.13.32=h3c776e5_2
aws-c-mqtt=0.9.5=h83b98fe_1
aws-c-s3=0.3.14=h24e141d_3
aws-c-sdkutils=0.1.12=he70778a_1
aws-checksums=0.1.17=he70778a_1
aws-crt-cpp=0.23.1=h04fc39a_0
aws-sdk-cpp=1.11.156=he404d52_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
backports.zoneinfo=0.2.1=py38hb991d35_7
blinker=1.6.2=pyhd8ed1ab_0
brotlipy=0.7.0=py38hb991d35_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.19.1=hb547adb_0
ca-certificates=2023.7.22=hf0a4a13_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38ha45ccd6_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
cloudpickle=2.0.0=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
cryptography=41.0.3=py38h92a0862_0
debugpy=1.7.0=py38he333c0f_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.12.3=pyhd8ed1ab_0
freetype=2.12.1=hd633e50_1
gflags=2.2.2=hc88da5d_1004
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.35=pyhd8ed1ab_0
glog=0.6.0=h6da1cb0_0
gmp=6.2.1=h9f76cd9_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.0.1=pyhd8ed1ab_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.12.2=pyhd1c38e8_0
ipywidgets=8.1.0=pyhd8ed1ab_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jmespath=1.0.1=pypi_0
jsonschema=4.19.0=pyhd8ed1ab_1
jsonschema-specifications=2023.7.1=pyhd8ed1ab_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.1=py38h10201cd_0
jupyterlab_widgets=3.0.8=pyhd8ed1ab_0
krb5=1.21.2=h92f50d5_0
lcms2=2.15=hd835a16_1
lerc=4.0.0=h9a09cb3_0
libabseil=20230125.3=cxx17_h13dd4ca_0
libarrow=13.0.0=h43f999d_3_cpu
libblas=3.9.0=18_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_0
libbrotlidec=1.1.0=hb547adb_0
libbrotlienc=1.1.0=hb547adb_0
libcblas=3.9.0=18_osxarm64_openblas
libcrc32c=1.1.2=hbdafb3b_0
libcurl=8.2.1=hc52a3a8_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.18=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libevent=2.1.12=h2757513_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libgoogle-cloud=2.12.0=h05652e3_1
libgrpc=1.56.2=h9075ed4_1
libjpeg-turbo=2.1.5.1=h1a8c8d9_0
liblapack=3.9.0=18_osxarm64_openblas
libnghttp2=1.52.0=hae82a92_0
libopenblas=0.3.24=openmp_hd76b1f2_0
libpng=1.6.39=h76d750c_0
libprotobuf=4.23.3=hf32f9b9_1
libsodium=1.0.18=h27ca646_1
libsqlite=3.43.0=hb31c410_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.19.0=ha061701_0
libtiff=4.5.1=h23a1a89_1
libutf8proc=2.8.0=h1a8c8d9_0
libwebp-base=1.3.1=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=16.0.6=h1c12783_0
lz4-c=1.9.4=hb7217d7_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.3=py38hb192615_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
ncurses=6.4=h7ea286d_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.24.4=py38ha84db1f_0
openjpeg=2.5.0=hbc2ba62_2
openssl=3.1.2=h53f4e23_0
orc=1.9.0=h858f345_1
oscrypto=1.2.1=pyhd3deb0d_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py38hefb543e_1
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=py38h3f590de_1
pip=23.2.1=pyhd8ed1ab_0
pipreq=0.4=pypi_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=3.8.1=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.23.3=py38hfca1a1f_0
psutil=5.9.5=py38hb991d35_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=13.0.0=py38h9b35334_3_cpu
pycparser=2.21=pyhd8ed1ab_0
pycryptodomex=3.18.0=py38hb192615_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pympler=1.0.1=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pysocks=1.7.1=pyha2e5f31_6
python=3.8.17=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.8=3_cp38
pytz=2023.3.post1=pyhd8ed1ab_0
pytz-deprecation-shim=0.1.0.post0=py38h10201cd_3
pyyaml=6.0.1=py38hb192615_0
pyzmq=25.1.1=py38hef91016_0
re2=2023.03.02=hc5e2d97_0
readline=8.2=h92ec313_1
referencing=0.30.2=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
rich=13.5.1=pyhd8ed1ab_0
rpds-py=0.10.2=py38hd0c8013_0
s3transfer=0.6.2=pypi_0
setuptools=68.1.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.10=h17c5cce_0
snowflake-connector-python=3.2.0=py38hafd26b6_0
snowflake-snowpark-python=1.6.1=pyhd8ed1ab_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.26.0=pyhd8ed1ab_0
tenacity=8.2.3=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
toml=0.10.2=pyhd8ed1ab_0
tomlkit=0.12.1=pyha770c72_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py38hb192615_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
tzlocal=4.3=py38h10201cd_0
urllib3=1.26.15=pyhd8ed1ab_0
utils=1.0.1=pypi_0
validators=0.22.0=pyhd8ed1ab_0
watchdog=3.0.0=py38hb991d35_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
widgetsnbextension=4.0.8=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.16.2=pyhd8ed1ab_0
zstd=1.5.5=h4f39d0f_0